// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
	/** A (potentially binary) string encoded using base64. */
	Base64String: any;
	/** An ISO-8601 encoded date string. */
	Date: any;
	/** An ISO-8601 encoded UTC date string. */
	DateTime: any;
	/** A Git object ID. */
	GitObjectID: any;
	/** A fully qualified reference name (e.g. `refs/heads/master`). */
	GitRefname: any;
	/** Git SSH string */
	GitSSHRemote: any;
	/** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
	GitTimestamp: any;
	/** A string containing HTML code. */
	HTML: any;
	/** An ISO-8601 encoded UTC date string with millisecond precision. */
	PreciseDateTime: any;
	/** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
	URI: any;
	/** A valid x509 certificate string */
	X509Certificate: any;
};
// ...
export type RepositoriesQueryVariables = Exact<{
	query?: Maybe<Scalars['String']>;
	after?: Maybe<Scalars['String']>;
	before?: Maybe<Scalars['String']>;
	first?: Maybe<Scalars['Int']>;
}>;